implement 404 error page and use usecontext hook
