+++ frontend/src/App.js
@@ -1,28 +1,27 @@
-import Navbar from './Components/Navbar';
+import Home from './Components/Home';
 import Footer from './Components/Footer';
 import Form from './Components/Form';
 import './App.css';
 import { Route, Routes } from 'react-router-dom';
-import Login from './Components/Login';
-import Signup from './Components/Signup';
 import Landingpage from './Components/Landingpage';
+import NoteState from './Context/NoteState';
 import Data from './Components/Data';
+import NotFoundPage from './Components/NotFoundpage';
 
 function App() {
   return (
     <div className="App">
-      {/* <ImageCarousel/> */}
-     <Navbar/>
+     <NoteState>
       <Routes>
+        <Route path='/home' element={<Home/>} />
         <Route path='/' element={<Landingpage />} />
-        <Route path='/signup' element={<Signup/>} />
-        <Route path='/login' element={<Login/>} />
-        
-        <Route path='/home' element={<></>} />
         <Route path='/form' element={<Form/>} />
         <Route path='/home/:id' element={<Data/>} />
+        <Route path='*' element={<NotFoundPage/>} />
       </Routes>
+     
       <Footer/>
+      </NoteState>
       
       
     </div>
