+++ frontend/src/Components/Footer.js
@@ -4,10 +4,11 @@ import { FaLinkedin,FaInstagram  } from "react-icons/fa";
 
 
 function Footer() {
+  
   return (
     <>
      <footer>
-      
+     
       <ul className='social_icon'>
         <li><a href='https://www.linkedin.com/in/anmol-singh-159734255/'><FaLinkedin style={{color:'#0072b1'}}/></a></li>
         <li><a href='https://www.instagram.com/anmol_virk0005/'><FaInstagram style={{color:'#fa7e1e'}}/></a></li>
