+++ frontend/src/Context/noteContext.js
@@ -1,0 +1,6 @@
+import { createContext } from "react";
+
+
+const noteContext=createContext();
+
+export default noteContext;
